finish implementing learning algorithm

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Progress.java
@@ -65,7 +65,7 @@ public class Progress {
         this.id = id;
     }
 
-    public long getEasinessFactor() {
+    public int getEasinessFactor() {
         return easinessFactor;
     }
